// backend/src/database/tables.d.ts
// This file is auto-generated by running "npm run db:dump-types"
// Do not edit this file directly

export type DatabaseMap = {
  users:
    | 'id'
    | 'username'
    | 'display_name'
    | 'email'
    | 'password'
    | 'avatar'
    | 'favorite_color'
    | 'created_at'
    | 'updated_at'
    | 'deleted_at'
    | 'github_id'
    | 'google_id';
  documents:
    | 'id'
    | 'title'
    | 'content'
    | 'binary_content'
    | 'created_at'
    | 'updated_at'
    | 'deleted_at';
  password_recoveries:
    | 'id'
    | 'user_id'
    | 'code'
    | 'expires_at'
    | 'created_at'
    | 'updated_at'
    | 'deleted_at';
  magic_links:
    | 'id'
    | 'user_id'
    | 'code'
    | 'expires_at'
    | 'created_at'
    | 'updated_at'
    | 'deleted_at';
};

import { Knex } from 'knex';